requirements.txt
fastapi==0.79.0

PyJWT==2.4.0
python-decouple==3.6
pymongo==4.2.0
dnspython==2.2.1
python-multipart==0.0.5

bcrypt==3.2.2
boto3==1.24.46
requests==2.28.1
beautifulsoup4==4.11.1

httptools==0.4.0
websockets==10.3
PyYAML==6.0
python-dotenv==0.20.0
watchfiles==0.16.1


uvicorn==0.18.2
pylint==2.14.5
autopep8==1.6.0
